implement and test writing to large files
The maximum supported file size is currently 2 MB, due to the current implementation.

When a block for a regular file is allocated in Storage.allocateFile,
its block size is 4096. Of these, 24 bytes are used for other data,
leaving 4072 bytes for the chunk references. Each chunk reference takes
8 bytes, which amounts to 509 chunk references.

Each chunk can hold 4096 bytes of payload data. The maximum size of a
file is therefore 509 * 4096 = 2084864 bytes.

This is too little for being actually useful. This limit can be raised
by changing the magic numbers in the Storage class, but this should not
be necessary. It would be better if the filesystem handled this
situation itself, which is entirely possible according to the
specification, it just hasn't been implemented yet.