fix real world test
This test still failed with the message containing 337, but only of pure
incidence. One of the larger files is Directory.class, which is 7433
bytes. Subtracting the chunk size 4096 results in 3337, and this matches
the (no longer) expected "337".

I should have written "expected: 337", but the chances of this test
failing in exactly this manner were really low, somewhere in the range
of 1/100 to 1/1000 since the string "337" could have also occurred in
the middle of the actual bytes read.