fix moving an object from one directory to the same directory
When a file is not really moved but only renamed, the old and new parent
directory are the same, and the two variables in Filesystem.move alias
each other. Therefore, when the new directory gets a new block number
because it was resized, the reference to the old directory must be
updated as well.