specify Gradle version
The previous Qodana run failed with:
> Cannot resolve constructor 'IndexOutOfBoundsException(long)'

In the previous commit, I had added the java.toolchain.languageVersion
in order to prevent exactly this compile-time error. The most plausible
explanation for now is that Qodana uses a very old Gradle version if no
version is specified in the Gradle wrapper directory. Rumor has it that
java.toolchain.languageVersion has been added in Gradle 5, which means
that the default Gradle version must be even older.

When Gradle downloads the correct version, it logs the download URL. Due
to another bug in IDEA, this Gradle log is truncated to 80 characters
per line when Qodana is run via the GitHub Action, which makes the
download log useless. The log contains:

    (Project configuration)
    Downloading https://services.gradle.org/distributions/g

To work around all these issues, try to specify the Gradle version to be
used.